Move allocator and deleter to own header
