+++ src/mjolnir/core/memory/linear_memory.h
@@ -13,6 +13,8 @@
 #include "mjolnir/core/exception.h"
 #include "mjolnir/core/fundamental_types.h"
 #include "mjolnir/core/memory/definitions.h"
+#include "mjolnir/core/memory/memory_system_allocator.h"
+#include "mjolnir/core/memory/memory_system_deleter.h"
 #include "mjolnir/core/memory/utility.h"
 #include "mjolnir/core/utility/pointer_operations.h"
 
@@ -23,11 +25,6 @@
 
 namespace mjolnir
 {
-template <typename, MemorySystem>
-class MemorySystemAllocator;
-template <typename, MemorySystem>
-class MemorySystemDeleter;
-
 // --- LinearMemory ---------------------------------------------------------------------------------------------------
 
 //! \addtogroup core_memory
@@ -290,185 +287,6 @@ private:
 };
 
 
-// --- MemorySystemAllocator ------------------------------------------------------------------------------------------
-
-//! @brief
-//! STL compatible allocator for memory systems.
-//!
-//! @tparam T_Type:
-//! Type of the allocated object
-//! @tparam T_MemorySystem:
-//! The memory system type.
-template <typename T_Type, MemorySystem T_MemorySystem>
-class MemorySystemAllocator
-{
-public:
-    //! @brief
-    //! The Type of the objects that are allocated by the allocator.
-    using ValueType = T_Type;
-
-
-    //! \cond DO_NOT_DOCUMENT
-
-    //! Required class members for allocators
-    using value_type = T_Type; // NOLINT(readability-identifier-naming)
-
-    template <typename T_OtherType>
-    struct rebind // NOLINT(readability-identifier-naming)
-    {
-        using other = MemorySystemAllocator<T_OtherType, T_MemorySystem>; // NOLINT(readability-identifier-naming)
-    };
-
-
-    MemorySystemAllocator()                                 = delete;
-    MemorySystemAllocator(const MemorySystemAllocator&)     = default;
-    MemorySystemAllocator(MemorySystemAllocator&&) noexcept = default;
-    ~MemorySystemAllocator()                                = default;
-    auto operator=(const MemorySystemAllocator&) -> MemorySystemAllocator& = default;
-    auto operator=(MemorySystemAllocator&&) noexcept -> MemorySystemAllocator& = default;
-    //! \endcond
-
-
-    //! @brief
-    //! Construct a new allocator with the passes `LinearMemory` instance
-    //!
-    //! @param[in] linear_memory:
-    //! `LinearMemory` instance that provides the memory for the allocations
-    explicit MemorySystemAllocator(T_MemorySystem& linear_memory) noexcept;
-
-
-    //! @brief
-    //! Construct a new allocator using the same `LinearMemory` instance as the passed allocator.
-    //!
-    //! @tparam T_OtherType
-    //! Object type of the other allocator
-    //!
-    //! @param[in] other:
-    //! Other allocator that provides a reference to the `LinearMemory` instance that should be used
-    template <class T_OtherType>
-    explicit MemorySystemAllocator(const MemorySystemAllocator<T_OtherType, T_MemorySystem>& other) noexcept;
-
-
-    //! @brief
-    //! Allocate memory for one or more instances of the allocators `T_Type`.
-    //!
-    //! @param[in] num_instances:
-    //! Specifies the number of instances of `T_Type` that should fit into the allocated memory
-    //!
-    //! @return
-    //! Pointer to the allocated memory
-    [[nodiscard]] auto allocate(UST num_instances) -> T_Type*;
-
-
-    //! @brief
-    //! Get an allocator with different value type that uses the same memory system.
-    //!
-    //! @tparam T_OtherType:
-    //! Type of the objects that should be allocated by the new allocator
-    //!
-    //! @return
-    //! New allocator that uses the same memory system
-    template <typename T_OtherType>
-    [[nodiscard]] auto as_type() const noexcept -> MemorySystemAllocator<T_OtherType, T_MemorySystem>;
-
-
-    //! @brief
-    //! Deallocates the memory of the passed pointer.
-    //!
-    //! @details
-    //! Due to the mechanics of the internally used `LinearMemory` this function does nothing.
-    //!
-    //! @param[in] pointer:
-    //! Pointer to the memory that should be freed
-    //! @param[in] num_instances:
-    //! Specifies how many instances of `T_Type` fitted into the allocated memory.
-    void deallocate([[maybe_unused]] T_Type* pointer, UST num_instances);
-
-
-    //! @brief
-    //! Get a reference to the memory system that is used by the allocator.
-    //!
-    //! @return
-    //! Memory system that is used by the allocator
-    [[nodiscard]] auto get_memory_system() const noexcept -> T_MemorySystem&;
-
-
-private:
-    T_MemorySystem& m_memory;
-
-    template <typename, MemorySystem>
-    friend class MemorySystemAllocator;
-};
-
-
-// --- MemorySystemDeleter --------------------------------------------------------------------------------------------
-
-//! @brief
-//! STL compatible deleter for memory systems.
-//!
-//! @tparam T_Type:
-//! Type of the objects that are deleted by this class
-//! @tparam T_MemorySystem:
-//! The memory system type.
-template <typename T_Type, MemorySystem T_MemorySystem>
-class MemorySystemDeleter
-{
-public:
-    //! @brief
-    //! The Type of the objects that are deleted by the deleter.
-    using ValueType = T_Type;
-
-    //! \cond DO_NOT_DOCUMENT
-    MemorySystemDeleter()                               = delete;
-    MemorySystemDeleter(const MemorySystemDeleter&)     = default;
-    MemorySystemDeleter(MemorySystemDeleter&&) noexcept = default;
-    ~MemorySystemDeleter()                              = default;
-    auto operator=(const MemorySystemDeleter&) -> MemorySystemDeleter& = default;
-    auto operator=(MemorySystemDeleter&&) noexcept -> MemorySystemDeleter& = default;
-    //! \endcond
-
-
-    //! @brief
-    //! Construct a new deleter with the passes `LinearMemory` instance.
-    //!
-    //! @param[in] memory_system:
-    //! Memory system that provided the memory for the object that should be deleted
-    explicit MemorySystemDeleter(T_MemorySystem& memory_system) noexcept;
-
-
-    //! @brief
-    //! Destroy the object at the passed memory address and deallocate the memory.
-    //!
-    //! @param[in] pointer:
-    //! Pointer to the object that should be destroyed and the memory that should be deallocated.
-    void operator()(std::remove_extent_t<T_Type>* pointer) noexcept;
-
-
-    //! @brief
-    //! Get a deleter for another type that uses the same memory system.
-    //!
-    //! @tparam T_OtherType:
-    //! Type of the objects that should be deleted by the new deleter
-    //!
-    //! @return
-    //! New deleter that uses the same memory system
-    template <typename T_OtherType>
-    [[nodiscard]] auto as_type() const noexcept -> MemorySystemDeleter<T_OtherType, T_MemorySystem>;
-
-
-    //! @brief
-    //! Get a reference to the memory system that is used by the deleter
-    //!
-    //! @return
-    //! Memory system that is used by the deleter
-    [[nodiscard]] auto get_memory_system() const noexcept -> T_MemorySystem&;
-
-
-private:
-    T_MemorySystem& m_memory;
-};
-
-
 //! @}
 } // namespace mjolnir
 
@@ -689,103 +507,6 @@ template <typename T_Lock, typename T_Deleter>
 auto LinearMemory<T_Lock, T_Deleter>::get_start_address() const noexcept -> UPT
 {
     return pointer_to_integer(m_memory.get());
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-MemorySystemAllocator<T_Type, T_MemorySystem>::MemorySystemAllocator(T_MemorySystem& linear_memory) noexcept
-    : m_memory{linear_memory}
-{
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-template <class T_OtherType>
-MemorySystemAllocator<T_Type, T_MemorySystem>::MemorySystemAllocator(
-        const MemorySystemAllocator<T_OtherType, T_MemorySystem>& other) noexcept
-    : m_memory{other.m_memory}
-{
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-[[nodiscard]] auto MemorySystemAllocator<T_Type, T_MemorySystem>::allocate(UST num_instances) -> T_Type*
-{
-    return static_cast<T_Type*>(m_memory.allocate(num_instances * sizeof(T_Type), alignof(T_Type)));
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-template <typename T_OtherType>
-[[nodiscard]] auto MemorySystemAllocator<T_Type, T_MemorySystem>::as_type() const noexcept
-        -> MemorySystemAllocator<T_OtherType, T_MemorySystem>
-{
-    return MemorySystemAllocator<T_OtherType, T_MemorySystem>(m_memory);
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-void MemorySystemAllocator<T_Type, T_MemorySystem>::deallocate(T_Type* pointer, UST num_instances)
-{
-    m_memory.deallocate(pointer, num_instances * sizeof(T_Type), alignof(T_Type));
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-[[nodiscard]] auto MemorySystemAllocator<T_Type, T_MemorySystem>::get_memory_system() const noexcept -> T_MemorySystem&
-{
-    return m_memory;
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-// cppcheck-suppress constParameter
-MemorySystemDeleter<T_Type, T_MemorySystem>::MemorySystemDeleter(T_MemorySystem& memory_system) noexcept
-    : m_memory(memory_system)
-{
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-void MemorySystemDeleter<T_Type, T_MemorySystem>::operator()(std::remove_extent_t<T_Type>* pointer) noexcept
-{
-    m_memory.destroy_deallocate(pointer);
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-template <typename T_OtherType>
-[[nodiscard]] auto MemorySystemDeleter<T_Type, T_MemorySystem>::as_type() const noexcept
-        -> MemorySystemDeleter<T_OtherType, T_MemorySystem>
-{
-    return MemorySystemDeleter<T_OtherType, T_MemorySystem>(m_memory);
-}
-
-
-// --------------------------------------------------------------------------------------------------------------------
-
-template <typename T_Type, MemorySystem T_MemorySystem>
-[[nodiscard]] auto MemorySystemDeleter<T_Type, T_MemorySystem>::get_memory_system() const noexcept -> T_MemorySystem&
-{
-    return m_memory;
 }
